signed char -> float
Former-commit-id: e335352916040711f1c58b4bb3320f9d1b606b8e [formerly c7d948dda54215c25fe75f014dc7875f52eb0f85]
Former-commit-id: 7e9278fb5726b9f51876fd00010e145b437f248d

+++ Plugins/Voxel/Source/Voxel/Classes/FlatWorldGenerator.h
@@ -35,7 +35,7 @@ class VOXEL_API UFlatWorldGenerator : public UVoxelWorldGenerator
 	GENERATED_BODY()
 
 public:
-	int GetDefaultValue(FIntVector Position);
+	float GetDefaultValue(FIntVector Position);
 	FColor GetDefaultColor(FIntVector Position);
 
 	// Height of the difference between full and empty
